Menu

+++ java01/ex00/User.java
@@ -2,15 +2,15 @@
 public class User {
 	private final int identifier;
 	private String name;
-	private double balance;
+	private long balance;
 
-	public User(int identifier, String name, double balance) {
+	public User(int identifier, String name, long balance) {
 		this.name = name;
 		this.identifier = identifier;
 		setBalance(balance);
 	}
 
-	public double setBalance(double balance) {
+	public long setBalance(long balance) {
 		if (balance < 0) {
 			System.out.println("User balance cannot be negative");
 			return -1;
@@ -23,7 +23,7 @@ public class User {
 		this.name = name;
 	}
 
-	public double getBalance() {
+	public long getBalance() {
 		return balance;
 	}
 
@@ -37,6 +37,6 @@ public class User {
 
 	@Override
     public String toString() {
-        return "User[ID: %d, Name: %s, Balance: %f]".formatted(identifier, name, balance);
+        return "User[ID: %d, Name: %s, Balance: %d]".formatted(identifier, name, balance);
     }
 }
